+++ src/main/java/insane96mcp/runeenchanting/setup/RERunes.java
@@ -141,6 +141,7 @@ public class RERunes {
     // All durability items
     public static final DeferredHolder<Rune, EnduringRune> ENDURING = register("enduring", new EnduringRune());
     public static final DeferredHolder<Rune, AtmosphericRune> ATMOSPHERIC = register("atmospheric", new AtmosphericRune());
+    public static final DeferredHolder<Rune, LapisFueledRune> LAPIS_FUELED = register("lapis_fueled", new LapisFueledRune());
 
     // Curse
     public static final DeferredHolder<Rune, CurseOfBindingRune> CURSE_OF_BINDING = register("curse_of_binding", new CurseOfBindingRune());